Unite implementation of `ImageSlice.withCGImage`
Using `_CGPixel`.
